Finished the Note App!
Finished video section 3: 23 "Requiring Arguments and Advanced Yargs"
The final step was to use Yargs to require that the necessary data,
that is the title and body, be provided when using an option, ex.
add, remove, that needs them. It also creates a --help flag that
displays the info on the different commands